move cli stuff to main.rs
