+++ project2/self_balancing_tree/src/main.rs
@@ -1,15 +1,11 @@
-
 mod avl;
 mod red_black;
-mod cli;
-
 
 use avl::AVLTree;
 use red_black::*;
-use cli::cli;
-
 
 use std::cell::RefCell;
+use std::io;
 use std::rc::Rc;
 
 fn test_avl_tree_josh(num: i32) {
@@ -236,7 +232,6 @@ fn test_tree_height() {
     let empty_root: AVLTree<i32> = AVLTree::Empty;
     println!("Non-empty: {:?}", root.tree_height());
     println!("Empty: {:?}", empty_root.tree_height())
-
 }
 
 fn test_insert_node() {
@@ -289,7 +284,7 @@ pub fn test_insert_node_all(num: i32) {
                 right_child: RefCell::new(Rc::new(AVLTree::Empty)),
                 height: RefCell::new(3),
             };
-        
+
             let rc_root = Rc::new(n4);
             println!("Enum Root: {:#?}", &rc_root);
             rc_root.print_inorder();
@@ -318,7 +313,7 @@ pub fn test_insert_node_all(num: i32) {
                 right_child: RefCell::new(Rc::new(AVLTree::Empty)),
                 height: RefCell::new(2),
             };
-            
+
             let rc_root = Rc::new(n5);
             println!("Enum Root: {:#?}", &rc_root);
             rc_root.print_inorder();
@@ -409,8 +404,40 @@ fn test_avl_search() {
     println!("Should be false: {}", n4.search_node(&100));
 }
 
-fn main() {
+pub fn cli() {
+    println!("1. Start AVL tree: init avl");
+    println!("2. Start Red-Black tree: init rb");
+    println!("3. Exit: exit");
 
+    let mut input = String::new();
+    io::stdin()
+        .read_line(&mut input)
+        .expect("Failed to read input.");
+
+    match input.trim() {
+        "init avl" => {
+            avl_cli();
+        }
+        "init rb" => {
+            rb_cli();
+        }
+        "exit" => return,
+        _ => {
+            println!("Invalid input");
+        }
+    }
+}
+
+fn avl_cli() {
+    println!("Executing avl cli");
+    let empty_root: AVLTree<i32> = AVLTree::Empty;
+}
+
+fn rb_cli() {
+    println!("Executing rb cli");
+}
+
+fn main() {
     //////// AVL TREE /////////////
     // test_avl_tree_josh(2);
     test_is_tree_empty();
@@ -420,14 +447,11 @@ fn main() {
 
     ////// END AVL TREE ///////////
 
-
     //////// RED BLACK TREE /////////////
 
     println!("Hello, world!");
 
     /////////////////////////////////////
-
-
 
     let mut rbt2 = RedBlackTree::new();
     rbt2.insert(8);
@@ -470,10 +494,7 @@ fn main() {
     // rbt2.insert(127);
 
     // println!("\n\n{:#?}", rbt2);
-    
 
-
-    
     // println!("Leaf nodes: {}", rbt2.count_leaves());
     // println!("Tree height: {}", rbt2.get_height());
     // rbt2.in_order_traversal();
@@ -483,9 +504,7 @@ fn main() {
     println!("\n\n\nSEARCH RESULT:");
     rbt2.search(8).display_tree();
 
-
     //////// END RED BLACK TREE /////////
-    
-    cli();
 
+    cli();
 }
